pyproject.toml
[project]
name = "reviv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.13",
    "cryptography>=46.0.3",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-storages>=1.14.6",
    "fido2>=2.0.0",
    "google-genai>=1.56.0",
    "jwt>=1.4.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
